--- line_of_credit_api/model traning /model_loader.py
@@ -1,3 +1,0 @@
-from model_def import ImprovedLineOfCreditApprovalModel
-
-model = ImprovedLineOfCreditApprovalModel.load("loc_approval_model.pkl")
