--- line_of_credit_api/requirements.txt
@@ -1,7 +1,0 @@
-fastapi
-uvicorn
-scikit-learn
-pandas
-numpy
-pydantic
-python-multipart
