+++ model traning /model_loader.py
@@ -1,0 +1,3 @@
+from model_def import ImprovedLineOfCreditApprovalModel
+
+model = ImprovedLineOfCreditApprovalModel.load("loc_approval_model.pkl")
